requirements.txt
numpy==1.26.4
Pillow==10.4.0
streamlit==1.36.0
tensorflow==2.17.0